removed reconvertion float64

+++ calculations/guessRange.go
@@ -3,8 +3,8 @@ package calculations
 import "math"
 
 func GuessRange(num []float64, currentNum float64, nextNum float64) (int, int) {
-	mean := float64(Mean(num))
-	standardDev := float64(StandardDev(num))
+	mean := Mean(num)
+	standardDev := StandardDev(num)
 
 	rangeStart := math.Max(currentNum-standardDev, mean-standardDev)
 	rangeEnd := math.Min(nextNum+standardDev, mean+standardDev)
